AccountService, AccountResource, AccountServiceTest, AccountResourceTest added.

+++ service/src/main/java/edu/lmu/cs/diabolical/ws/dao/AccountDaoHibernateImpl.java
@@ -9,7 +9,7 @@ import edu.lmu.cs.diabolical.ws.domain.Account;
 public class AccountDaoHibernateImpl extends HibernateDaoSupport implements AccountDao{
 
     @Override
-    public Account findAccountById(Integer id) {
+    public Account findAccountById(Long id) {
         return getHibernateTemplate().get(Account.class, id);
     }
